+++ .byebug_history
@@ -1,4 +1,9 @@
 c
+params[:activities]
+params[:activities].is_a?(Array)
+c
+params[:activities]
+c
 activities
 c
 activity_id
